Created GreetingController.java
This seems to initialize the template for the 'content' string for the 200 response to the /greeting endpoint.

Seems like this controller allows a new greetingObject to be created with the concept being the counter setting the 'id' for the Greeting object and the String template (and defaultValue passed in  for line 17 (that I don't quite completely conceptually understand) to make an instance of the Greeting Object with some values of some sort in place for it's attributes.